refactor: move covariance modification functions to own module
